requirements.txt
pynacl>=1.5.0
base58>=2.1.0
requests>=2.28.0
